// frontend/src/app/models/cliente.model.ts
export interface Cliente {
  ID: number;
  idUsuario: number;
  DataHoraCadastro: string;
  Codigo: string;
  Nome: string;
  CPF_CNPJ: string;
  CEP: number;
  Logradouro: string;
  Endereco: string;
  Numero: string;
  Bairro: string;
  Cidade: string;
  UF: string;
  Complemento: string;
  Fone: string;
  LimiteCredito: number;
  Validade: string;
}